+++ src/main/java/org/halvors/nuclearphysics/common/block/BlockRadioactive.java
@@ -84,7 +84,7 @@ public class BlockRadioactive extends BlockBase {
     }
 
     /**
-     * Called whenever an entity is walking on top of this block. Args: world, x, y, z, entity
+     * Called whenever an entity is walking on top of this block. Args: worldgen, x, y, z, entity
      */
     @Override
     public void onEntityWalking(final World world, final int x, final int y, final int z, final Entity entity) {
