+++ src/main/java/org/halvors/nuclearphysics/common/init/ModWorldGenerators.java
@@ -1,10 +1,10 @@
 package org.halvors.nuclearphysics.common.init;
 
 import cpw.mods.fml.common.registry.GameRegistry;
-import org.halvors.nuclearphysics.common.world.WorldGeneratorOre;
+import org.halvors.nuclearphysics.common.worldgen.OreGenerator;
 
 public class ModWorldGenerators {
     public static void registerWorldGenerators() {
-        GameRegistry.registerWorldGenerator(new WorldGeneratorOre(), 0);
+        GameRegistry.registerWorldGenerator(new OreGenerator(), 0);
     }
 }
